# src/main/resources/openapi3.yml
openapi: 3.0.1
info:
  title: Open Data Hub Mobility API
  description: >
    Documentation: <a
    href="https://opendatahub.readthedocs.io/en/latest/datasets.html#datasets-in-the-mobility-domain">Mobility
    Datasets</a> |
    <a
    href="https://opendatahub.readthedocs.io/en/latest/howto/mobility/getstarted.html">Getting
    started with the API Mobility V2</a> | <a
    href="https://opendatahub.readthedocs.io/en/latest/howto.html">General
    Howtos</a> | <a
    href="https://github.com/noi-techpark/bdp-core">Source Code</a>

  termsOfService: https://opendatahub.readthedocs.io/en/latest/licenses.html#apis-terms-of-service
  contact:
    name: Open Data Hub
    url: https://opendatahub.bz.it
    email: help@opendatahub.bz.it
  license:
    name: API License
    url: https://opendatahub.readthedocs.io/en/latest/licenses.html
  version: V1 and V2
servers:
  - url: __ODH_SERVER_URL__
...
paths:
...
  /v2/{representation}/{stationTypes}/{dataTypes}/latest:
    get:
      tags:
        - Mobility V2
      summary:
        View details of all given station types including data types and most recent
        measurements
      description:
        You can put multiple stationtypes or datatypes as comma-seperated lists.
        The response contains stations (prefix <code>s</code>), their
        parent-stations (prefix <code>p</code>), types (prefix <code>t</code>)
        and the latest measurements (prefix <code>m</code>). Each JSON attribute
        is prefixed with the first letter of that object.
      parameters:
        - $ref: "#/components/parameters/representation"
        - $ref: "#/components/parameters/stationTypes"
        - $ref: "#/components/parameters/dataTypes"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/select"
        - $ref: "#/components/parameters/where"
        - $ref: "#/components/parameters/shownull"
        - $ref: "#/components/parameters/distinct"
        - $ref: "#/components/parameters/timezone"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
...
components:
...
  parameters:
...
    representation:
      name: representation
      in: path
      description:
        Do you want to have the result in a <code>tree</code> or
        <code>flat</code> representation, and do you want to get
        <code>node</code> (=station), <code>edge</code> or <code>event</code>
        information. Add them as comma-separated values.
      required: true
      schema:
        type: string
        default: flat,node
...
    stationTypes:
      name: stationTypes
      in: path
      description:
        Station types or categories. Multiple types possible as comma-separated-values.
        All types with <code>*</code>.
      required: true
      schema:
        type: string
        default: "*"
...
    dataTypes:
      name: dataTypes
      in: path
      description:
        Data types. Multiple types possible as comma-separated-values.
        All types with <code>*</code>.
      required: true
      schema:
        type: string
        default: "*"
...
    limit:
      name: limit
      in: query
      description: The limit of the response. Set it to -1 to disable it.
      schema:
        type: integer
        format: int64
        default: 200
    offset:
      name: offset
      in: query
      description:
        The offset of the response list. To simulate pagination, together
        with limit.
      schema:
        type: integer
        format: int64
        default: 0
    timezone:
      name: timezone
      in: query
      description: Set an optional timezone for all response timestamps. See java.time.ZoneId for further information.
      schema:
        type: string
        default: UTC
    select:
      name: select
      in: query
      description: >
        Select <code>aliases</code>, which will be used to build the
        response. Multiple aliases possible as comma-separated-values. Example:
        <code>sname</code> or <code>smetadata.city.cap</code> for JSON. Functions
        can be set as <code>func(alias)</code> (Functions with JSON are not supported
        yet)
      schema:
        type: string
    where:
      name: where
      in: query
      description: |-
        Filter the result with filter-triples, like <code>alias.operator.value_or_list</code>

        <code>values_or_list</code>
         -   value: Whatever you want, also a regular expression. However, you need to escape <code>,'"</code> with a <code>\\</code>. Use url-encoded values, if your tool does not support certain characters.
         -   list: <code>(value,value,value)</code>

        <code>operator</code>
         -   eq: Equal
         -   neq: Not Equal
         -   lt: Less Than
         -   gt: Greater Than
         -   lteq: Less Than Or Equal
         -   gteq: Greater Than Or Equal
         -   re: Regular Expression
         -   ire: Insensitive Regular Expression
         -   nre: Negated Regular Expression
         -   nire: Negated Insensitive Regular Expression
         -   bbi: Bounding box intersecting objects (ex., a street that is only partially covered by the box). Syntax? See below.
         -   bbc: Bounding box containing objects (ex., a station or street, that is completely covered by the box). Syntax? See below.
         -   in: True, if the value of the alias can be found within the given list. Example: name.in.(Peter,Patrick,Rudi)
         -   nin: False, if the value of the alias can be found within the given list. Example: name.nin.(Peter,Patrick,Rudi)


        <code>logical operations</code>
         -   and(alias.operator.value_or_list,...): Conjunction of filters (can be nested)
         -   or(alias.operator.value_or_list,...): Disjunction of filters (can be nested)

        Multiple conditions possible as comma-separated-values. <code>value</code>s will be casted to Double precision or <code>null</code>, if possible. Put them inside double quotes, if you want to prevent that.

        Example-syntax for bbi/bbc could be <code>coordinate.bbi.(11,46,12,47,4326)</code>, where the ordering inside the list is left-x, left-y, right-x, right-y and SRID (optional).
      schema:
        type: string
    shownull:
      name: shownull
      in: query
      description:
        Should JSON keys with null-values be returned, or removed from
        the response-JSON.
      schema:
        type: boolean
        default: false
    distinct:
      name: distinct
      in: query
      description: Remove duplicate entries.
      schema:
        type: boolean
        default: true

    # V1 PARAMETERS -----------------------------------------------------------------------------
...